moved ListToolInforHandler to separate service, fixed doc strings, cleaned up comments
