+++ jupyter_server/services/contents/handlers.py
@@ -424,13 +424,6 @@ class TrustNotebooksHandler(JupyterHandler):
         self.finish()
 
 
-# Somehow this doesn't feel like the right service for this to go in?
-class ListToolInfoHandler(APIHandler):
-    @web.authenticated
-    async def get(self):
-        tools = self.serverapp.extension_manager.discover_tools()
-        self.finish({"discovered_tools": tools})
-
 
 # -----------------------------------------------------------------------------
 # URL to handler mappings
@@ -449,5 +442,4 @@ default_handlers = [
     (r"/api/contents%s/trust" % path_regex, TrustNotebooksHandler),
     (r"/api/contents%s" % path_regex, ContentsHandler),
     (r"/api/notebooks/?(.*)", NotebooksRedirectHandler),
-    (r"/api/tools", ListToolInfoHandler),
 ]
